TLRC - 895 - Implement Error Reporting for Failed Uploads (Stage 2 Validations) - TLRC - 1044 - Dev - Implemented complete Download Functionality

+++ src/Sfa.Tl.ResultsAndCertification.Web/ViewModel/Registration/UploadRegistrationsRequestViewModel.cs
@@ -13,7 +13,7 @@ namespace Sfa.Tl.ResultsAndCertification.Web.ViewModel.Registration
 {
     public class UploadRegistrationsRequestViewModel
     {
-        public int AoUkprn { get; set; }
+        public long AoUkprn { get; set; }
         public string BlobFileName { get; set; }
         public Guid BlobUniqueReference { get; set; }
         public FileType FileType { get; set; }
