+++ src/Sfa.Tl.ResultsAndCertification.Web/ViewModel/Registration/UploadRegistrationsResponseViewModel.cs
@@ -6,7 +6,7 @@ namespace Sfa.Tl.ResultsAndCertification.Web.ViewModel.Registration
     {
         public bool IsSuccess { get; set; }
         public Guid BlobUniqueReference { get; set; }
-        public long ErrorFileSize { get; set; }
+        public double ErrorFileSize { get; set; }
         public StatsViewModel Stats { get; set; }
     }
 }
